Nested sym table structure finished, not tested
The Scope subclases (Base, Local, Global) were not needed, since the MethodSymbol is always Local scope, and has Scope interface implementation and it's own sym table, it could be recursive and could be added to GLOBAL scope which is the SymbolsTable.
Also was added the push and pop Scope to the SymbolsTable, which will be the one that controlls all the Scope Stack with the currentScope variable.
If more symbols that has scope needs to be added (ex.if, while, for) just need to implement Scope interface, and has their own sym table, as the MethodSymbols is doing