Added 2 extra rules, fixed compiling errors, added reset.sh script, found the regex error to get return and a file CompileLog with the output.
The rules added was, one to enable var = operation; and other to enable Expresiones containing a block, which won't have another block inside, this rule couldn't be tested correctly since the regexp error found to get "return" statement.
The return statement is identified as VARIABLE instead of RETURN, sin ce the first one covers it with [A-Za-z]+

The reset script clean the bytecode files, recompile all the project and run it.